fix:moved item block render into new component
